refactor in preperation for adding gues functionality
