+++ frontend/containers/Host/EventDetail.js
