+++ frontend/containers/Root/Root.dev.js
@@ -1,12 +1,16 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { Provider } from 'react-redux';
-import AppContainer from './AppContainer.js';
+import {Provider} from 'react-redux';
+import AppContainer from '../AppContainer.js';
+import DevTools from './DevTools';
 
 export default function Root({ store }) {
     return (
         <Provider store={store}>
-            <AppContainer />
+            <div>
+                <AppContainer />
+                <DevTools />
+            </div>
         </Provider>
     );
 }
