+++ frontend/containers/Root/Root.prod.js
@@ -1,16 +1,12 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import {Provider} from 'react-redux';
-import AppContainer from './AppContainer.js';
-import DevTools from './DevTools';
+import { Provider } from 'react-redux';
+import AppContainer from '../AppContainer.js';
 
 export default function Root({ store }) {
     return (
         <Provider store={store}>
-            <div>
-                <AppContainer />
-                <DevTools />
-            </div>
+            <AppContainer />
         </Provider>
     );
 }
