+++ frontend/store/configureStore.dev.js
@@ -1,6 +1,6 @@
 import { createStore, compose, applyMiddleware } from 'redux';
 import reducers from '../reducers';
-import DevTools from '../containers/DevTools';
+import DevTools from '../containers/Root/DevTools';
 import reduxThunk from 'redux-thunk';
 
 export function configureStore(initialState) {
